+++ jscatter/widget.py
@@ -193,7 +193,7 @@ class JupyterScatter(anywidget.AnyWidget):
     connection_color = Union([Union([Unicode(), List(minlen=4, maxlen=4)]), List(Union([Unicode(), List(minlen=4, maxlen=4)]))]).tag(sync=True)
     connection_color_selected = Union([Union([Unicode(), List(minlen=4, maxlen=4)]), List(Union([Unicode(), List(minlen=4, maxlen=4)]))]).tag(sync=True)
     connection_color_hover = Union([Union([Unicode(), List(minlen=4, maxlen=4)]), List(Union([Unicode(), List(minlen=4, maxlen=4)]))]).tag(sync=True)
-    connection_color_by = Enum([None, 'valueA', 'valueB', 'segment'], allow_none=True, default_value=None).tag(sync=True)
+    connection_color_by = Enum([None, 'valueA', 'valueB', 'segment', 'inherit'], allow_none=True, default_value=None).tag(sync=True)
     connection_opacity = Union([Float(), List(Float())], allow_none=True).tag(sync=True)
     connection_opacity_by = Enum([None, 'valueA', 'valueB', 'segment'], allow_none=True, default_value=None).tag(sync=True)
     connection_size = Union([Union([Int(), Float()]), List(Union([Int(), Float()]))]).tag(sync=True)